requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.6
django-restframework==0.0.1
djangorestframework==3.14.0
djangorestframework-jwt==1.0.1
djangorestframework-simplejwt==5.3.0
idna==3.7
psycopg2-binary==2.9.7
pykakasi==2.2.1
python-dotenv==1.0.0
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
